Fix editor
There was a bug that when dragging the handlers in the editor, the input value did not change. It turned out that this was because when introducing the polyline and polygon the input changed too eagerly and resulted in bad behavor. Fixed it in way that now the poligon and polyline does not change syncronously with the store, but updates only a local state and gets synced with the store on submit. Store change also gets syncronized with the input right away.