Seemingly fix an annoying bug that happens scrollbars are present
It turns out the 100vw does not consider the scroll bar width, so when you try to have something with the width of 100vw it will automatically have a horizontal scrollbar because of the vertical scrollbar. This seemingly meaningless calc seems to solve the issue. At least on Mac. Let's see what happens on Windows.